fix(firmware): clear the clang-tidy baseline, and analyse the ARM build
The CI job ran clang-tidy over all 67 firmware sources using the *SITL*
compile database. Eleven of them are not in that database, so they were
analysed with default flags and reported undeclared-function and parse
errors that do not exist in the flight build.

Split it: one job over the ARM database (all 67 sources, via the new
tools/dev/run_clang_tidy.sh, whose --extra-args are each documented with
the false-positive class they remove -- without the newlib -isystem, five
files die on "'string.h' file not found" and are not analysed at all), and
one over the SITL database restricted to the files that database actually
contains. Both configurations are now at zero findings.

Two of the findings were real defects, not noise:

  * navlink_tx.c called angle_controller_height_state() with no prototype
    in scope -- the one new warning in an otherwise warning-clean flight
    build. Correct code by luck of the AAPCS, a constraint violation since
    C99, an error by default in GCC 14.
  * vertical_task.c reached vayu_log transitively.

The rest are widening fixes on index arithmetic, plus two places where the
check is answered with a comment rather than obeyed: pid_config.c (int8_t
holding +/-1, where the suggested unsigned round-trip would read -1 as 255)
and bmx160.c, which now documents that bmx160_acc_range_t's enumerators are
g-values while the fields hold register codes -- the trap that will make the
first attempt at configuring the IMU look like it worked.

attitude_task.c now derives the reported estimator rate from the decimation
instead of reporting the nominal target, which stops agreeing with reality
as soon as the sample rate does not divide evenly.

+++ firmware/src/control/pid_config.c
@@ -169,7 +169,9 @@ bool pid_config_get_motor_geometry(float pos_x[4], float pos_y[4],
       pos_y[i] = s_store.motor_pos_y[i];
     }
     if (spin) {
-      spin[i] = s_store.motor_spin[i];
+      /* int8_t holding +1/-1: the sign is the payload, so the check's
+       * suggested unsigned-char round-trip would invert -1 to 255. */
+      spin[i] = (int)s_store.motor_spin[i];
     }
   }
   return true;
